Ulitise bufferedQueue for informerProcessorListenere and add tests for it
